Add integration test for Take Part
Move the common setup and helper code out of Statistics Announcement
test, so it can be used by Take Part, and future format tests.

Infer the format from the test class name, so it doesn't need to be
manually set in each test, but allow it to be overriden as a method
if a test file's name doesn't match the convention.

The tests added for take part aren't amazing, but they ensure that
the integration helpers aren't linked to a single format, and
give a basic example to add to later, if we need to.
